Separate swap input and native inputs into hook
