+++ src/screens/ExchangeModal.js
@@ -1,5 +1,5 @@
 import analytics from '@segment/analytics-react-native';
-import { get, isNil } from 'lodash';
+import { get } from 'lodash';
 import PropTypes from 'prop-types';
 import React, {
   Fragment,
@@ -26,13 +26,6 @@ import { GasSpeedButton } from '../components/gas';
 import { Centered, KeyboardFixedOpenLayout } from '../components/layout';
 import ExchangeModalTypes from '../helpers/exchangeModalTypes';
 import {
-  convertAmountFromNativeValue,
-  convertAmountToNativeAmount,
-  greaterThanOrEqualTo,
-  isZero,
-  updatePrecisionToDisplay,
-} from '../helpers/utilities';
-import {
   useAccountSettings,
   useBlockPolling,
   useGas,
@@ -40,10 +33,10 @@ import {
   usePrevious,
   useSwapDetails,
   useSwapInputRefs,
+  useSwapInputs,
   useUniswapAllowances,
   useUniswapCurrencies,
   useUniswapMarketDetails,
-  useUniswapMarketPrice,
 } from '../hooks';
 import { loadWallet } from '../model/wallet';
 import { executeRap } from '../raps/common';
@@ -92,13 +85,7 @@ const ExchangeModal = ({
   const { web3ListenerInit, web3ListenerStop } = useBlockPolling();
   const { nativeCurrency } = useAccountSettings();
   const prevSelectedGasPrice = usePrevious(selectedGasPrice);
-  const { getMarketPrice } = useUniswapMarketPrice();
   const { getMarketDetails } = useUniswapMarketDetails();
-
-  const [isMax, setIsMax] = useState(false);
-  const [inputAmount, setInputAmount] = useState(null);
-  const [inputAmountDisplay, setInputAmountDisplay] = useState(null);
-  const [inputAsExactAmount, setInputAsExactAmount] = useState(true);
   const { inputBalance, updateInputBalance } = useMaxInputBalance();
 
   const {
@@ -108,8 +95,6 @@ const ExchangeModal = ({
   } = useSwapDetails();
 
   const [isAuthorizing, setIsAuthorizing] = useState(false);
-  const [isSufficientBalance, setIsSufficientBalance] = useState(true);
-  const [nativeAmount, setNativeAmount] = useState(null);
   const [outputAmount, setOutputAmount] = useState(null);
   const [outputAmountDisplay, setOutputAmountDisplay] = useState(null);
   const [showConfirmButton, setShowConfirmButton] = useState(
@@ -145,10 +130,32 @@ const ExchangeModal = ({
     isDeposit,
     isWithdrawal,
     navigation,
-    setInputAsExactAmount,
     setShowConfirmButton,
     type,
     underlyingPrice,
+  });
+
+  const {
+    inputAmount,
+    inputAmountDisplay,
+    inputAsExactAmount,
+    isMax,
+    isSufficientBalance,
+    nativeAmount,
+    setInputAsExactAmount,
+    setIsSufficientBalance,
+    updateInputAmount,
+    updateNativeAmount,
+  } = useSwapInputs({
+    defaultInputAsset,
+    inputBalance,
+    inputCurrency,
+    isDeposit,
+    isWithdrawal,
+    nativeFieldRef,
+    outputCurrency,
+    supplyBalanceUnderlying,
+    type,
   });
 
   const updateGasLimit = useCallback(
@@ -201,106 +208,6 @@ const ExchangeModal = ({
     updateGasLimit,
   ]);
 
-  const updateInputAmount = useCallback(
-    (
-      newInputAmount,
-      newAmountDisplay,
-      newInputAsExactAmount = true,
-      newIsMax = false
-    ) => {
-      setInputAmount(newInputAmount);
-      setInputAsExactAmount(newInputAsExactAmount);
-      setInputAmountDisplay(
-        newAmountDisplay !== undefined ? newAmountDisplay : newInputAmount
-      );
-      setIsMax(newInputAmount && newIsMax);
-
-      if (!nativeFieldRef.current.isFocused() || newIsMax) {
-        let newNativeAmount = null;
-
-        const isInputZero = isZero(newInputAmount);
-
-        if (newInputAmount && !isInputZero) {
-          let newNativePrice = get(inputCurrency, 'native.price.amount', null);
-          if (isNil(newNativePrice)) {
-            newNativePrice = getMarketPrice(inputCurrency, outputCurrency);
-          }
-          newNativeAmount = convertAmountToNativeAmount(
-            newInputAmount,
-            newNativePrice
-          );
-        }
-        setNativeAmount(newNativeAmount);
-
-        if (inputCurrency) {
-          const newIsSufficientBalance =
-            !newInputAmount ||
-            (isWithdrawal
-              ? greaterThanOrEqualTo(supplyBalanceUnderlying, newInputAmount)
-              : greaterThanOrEqualTo(inputBalance, newInputAmount));
-
-          setIsSufficientBalance(newIsSufficientBalance);
-        }
-      }
-
-      if (newAmountDisplay) {
-        analytics.track('Updated input amount', {
-          category: isDeposit ? 'savings' : 'swap',
-          defaultInputAsset: defaultInputAsset && defaultInputAsset.symbol,
-          type,
-          value: Number(newAmountDisplay.toString()),
-        });
-      }
-    },
-    [
-      defaultInputAsset,
-      getMarketPrice,
-      inputBalance,
-      inputCurrency,
-      isDeposit,
-      isWithdrawal,
-      nativeFieldRef,
-      outputCurrency,
-      supplyBalanceUnderlying,
-      type,
-    ]
-  );
-
-  const updateNativeAmount = useCallback(
-    nativeAmount => {
-      logger.log('update native amount', nativeAmount);
-      let inputAmount = null;
-      let inputAmountDisplay = null;
-
-      const isNativeZero = isZero(nativeAmount);
-      setNativeAmount(nativeAmount);
-
-      setIsMax(false);
-
-      if (nativeAmount && !isNativeZero) {
-        let nativePrice = get(inputCurrency, 'native.price.amount', null);
-        if (isNil(nativePrice)) {
-          nativePrice = getMarketPrice(inputCurrency, outputCurrency);
-        }
-        inputAmount = convertAmountFromNativeValue(
-          nativeAmount,
-          nativePrice,
-          inputCurrency.decimals
-        );
-        inputAmountDisplay = updatePrecisionToDisplay(
-          inputAmount,
-          nativePrice,
-          true
-        );
-      }
-
-      setInputAmount(inputAmount);
-      setInputAmountDisplay(inputAmountDisplay);
-      setInputAsExactAmount(true);
-    },
-    [getMarketPrice, inputCurrency, outputCurrency]
-  );
-
   const updateOutputAmount = useCallback(
     (newOutputAmount, newAmountDisplay, newInputAsExactAmount = false) => {
       setInputAsExactAmount(newInputAsExactAmount);
@@ -317,7 +224,7 @@ const ExchangeModal = ({
         });
       }
     },
-    [defaultInputAsset, isDeposit, isWithdrawal, type]
+    [defaultInputAsset, isDeposit, isWithdrawal, setInputAsExactAmount, type]
   );
 
   const clearForm = useCallback(() => {
@@ -465,6 +372,7 @@ const ExchangeModal = ({
     outputCurrency,
     outputFieldRef,
     updateExtraTradeDetails,
+    setIsSufficientBalance,
     updateInputAmount,
     updateOutputAmount,
   ]);
@@ -510,6 +418,7 @@ const ExchangeModal = ({
     outputCurrencyUniqueId,
     outputFieldRef,
     outputReserveTokenAddress,
+    setIsSufficientBalance,
     updateExtraTradeDetails,
     updateInputAmount,
     updateOutputAmount,
